+++ .drone/scripts/install-deps.sh
@@ -7,13 +7,16 @@ set -ex
 # Additionally, we use a single file to install deps instead of one for each CI
 # script as it makes things easier to manage.
 apt_update_cmd=('apt-get' 'update')
+apt_upgrade_cmd=('apt-get' 'upgrade' '-y')
 apt_install_cmd=('apt-get' 'install' 'asciidoctor' 'bats' 'curl' 'debhelper' 'git' 'gpg' 'jq' 'openssh-client' 'python3' 'python3-pip' 'python3-requests' 'sed' 'sudo' 'tzdata' 'ubuntu-dev-tools' '-y')
 
 if [[ -z "${NO_SUDO}" ]]; then
     sudo "${apt_update_cmd[@]}"
+    sudo "${apt_upgrade_cmd[@]}"
     sudo "${apt_install_cmd[@]}"
 else
     "${apt_update_cmd[@]}"
+    "${apt_upgrade_cmd[@]}"
     "${apt_install_cmd[@]}"
 fi